feat: criado metodo para carregar mostrar usuários por login e corrigido problemas

+++ src/main/java/dao/DAOUsuarioRepository.java
@@ -1,111 +1,178 @@
 package dao;
 
-import connection.SingleConnectionBanco;
-import model.ModelLogin;
-
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
+import connection.SingleConnectionBanco;
+import model.ModelLogin;
+
 public class DAOUsuarioRepository {
-
-   private Connection connection;
-
+    private Connection connection;
     public DAOUsuarioRepository() {
         connection = SingleConnectionBanco.getConnection();
     }
 
-    public ModelLogin gravarUsuario(ModelLogin modelLogin) throws Exception {
+    public ModelLogin gravarUsuario(ModelLogin objeto, Long userLogado) throws Exception {
 
-        if(modelLogin.isNovo()) {
-            String sql = "INSERT INTO model_login(login, senha, nome, email) VALUES (?, ?, ?, ?)";
+        if (objeto.isNovo()) {/*Grava um novo*/
 
+            String sql = "INSERT INTO model_login(login, senha, nome, email, usuario_id)  VALUES (?, ?, ?, ?, ?);";
             PreparedStatement preparedSql = connection.prepareStatement(sql);
-            preparedSql.setString(1, modelLogin.getLogin());
-            preparedSql.setString(2, modelLogin.getSenha());
-            preparedSql.setString(3, modelLogin.getNome());
-            preparedSql.setString(4, modelLogin.getEmail());
 
+            preparedSql.setString(1, objeto.getLogin());
+            preparedSql.setString(2, objeto.getSenha());
+            preparedSql.setString(3, objeto.getNome());
+            preparedSql.setString(4, objeto.getEmail());
+            preparedSql.setLong(5, userLogado);
             preparedSql.execute();
-            connection.commit();
-        } else {
-            String sql = "UPDATE model_login SET login=?, senha=?, nome=?, email=? WHERE id="+modelLogin.getId()+"";
-            PreparedStatement prepareSql = connection.prepareStatement(sql);
-            prepareSql.setString(1, modelLogin.getLogin());
-            prepareSql.setString(2, modelLogin.getSenha());
-            prepareSql.setString(3, modelLogin.getNome());
-            prepareSql.setString(4, modelLogin.getEmail());
 
+            connection.commit();
+
+        }else {
+            String sql = "UPDATE model_login SET login=?, senha=?, nome=?, email=?  WHERE id =  "+objeto.getId()+";";
+
+            PreparedStatement prepareSql = connection.prepareStatement(sql);
+
+            prepareSql.setString(1, objeto.getLogin());
+            prepareSql.setString(2, objeto.getSenha());
+            prepareSql.setString(3, objeto.getNome());
+            prepareSql.setString(4, objeto.getEmail());
             prepareSql.executeUpdate();
+
             connection.commit();
         }
-
-        return this.consultaUsuario(modelLogin.getLogin());
+        return this.consultaUsuario(objeto.getLogin(), userLogado);
     }
+    public List<ModelLogin> consultaUsuarioList(Long userLogado) throws Exception {
 
-    public List<ModelLogin> consultarUsuarioList(String nome) throws Exception{
         List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 
-        String sql = "SELECT * FROM model_login WHERE UPPER(nome) LIKE UPPER (?) ";
+        String sql = "select * from model_login where useradmin is false and usuario_id = " + userLogado;
         PreparedStatement statement = connection.prepareStatement(sql);
-        statement.setString(1, "%" + nome + "%");
-        ResultSet resultSet = statement.executeQuery();
 
-        return getModelLogins(retorno, resultSet);
-    }
+        ResultSet resultado = statement.executeQuery();
 
-    public List<ModelLogin> consultarUsuarioListAll() throws Exception{
-        List<ModelLogin> retorno = new ArrayList<ModelLogin>();
-
-        String sql = "SELECT * FROM model_login";
-        PreparedStatement statement = connection.prepareStatement(sql);
-        ResultSet resultSet = statement.executeQuery();
-
-        return getModelLogins(retorno, resultSet);
-    }
-
-    private List<ModelLogin> getModelLogins(List<ModelLogin> retorno, ResultSet resultSet) throws SQLException {
-        while (resultSet.next()) {
+        while (resultado.next()) { /*percorrer as linhas de resultado do SQL*/
 
             ModelLogin modelLogin = new ModelLogin();
-            modelLogin.setEmail(resultSet.getString("email"));
-            modelLogin.setId(resultSet.getLong("id"));
-            modelLogin.setLogin(resultSet.getString("login"));
-            modelLogin.setNome(resultSet.getString("nome"));
+
+            modelLogin.setEmail(resultado.getString("email"));
+            modelLogin.setId(resultado.getLong("id"));
+            modelLogin.setLogin(resultado.getString("login"));
+            modelLogin.setNome(resultado.getString("nome"));
 
             retorno.add(modelLogin);
-
         }
-
         return retorno;
     }
 
-    public ModelLogin consultaUsuario(String login) throws Exception{
-        ModelLogin modelLogin = new ModelLogin();
+    public List<ModelLogin> consultaUsuarioList(String nome, Long userLogado) throws Exception {
 
-        String sql = "SELECT * FROM model_login WHERE UPPER(login) = UPPER('"+login+"')";
+        List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 
+        String sql = "select * from model_login  where upper(nome) like upper(?) and useradmin is false and usuario_id = ?";
         PreparedStatement statement = connection.prepareStatement(sql);
-        return getModelLogin(modelLogin, statement);
+        statement.setString(1, "%" + nome + "%");
+        statement.setLong(2, userLogado);
+
+        ResultSet resultado = statement.executeQuery();
+
+        while (resultado.next()) { /*percorrer as linhas de resultado do SQL*/
+
+            ModelLogin modelLogin = new ModelLogin();
+
+            modelLogin.setEmail(resultado.getString("email"));
+            modelLogin.setId(resultado.getLong("id"));
+            modelLogin.setLogin(resultado.getString("login"));
+            modelLogin.setNome(resultado.getString("nome"));
+
+            retorno.add(modelLogin);
+        }
+        return retorno;
     }
 
-    public ModelLogin consultaUsuarioID(String id) throws Exception{
+    public ModelLogin consultaUsuarioLogado(String login) throws Exception  {
+
         ModelLogin modelLogin = new ModelLogin();
 
-        String sql = "SELECT * FROM model_login WHERE id = ?";
+        String sql = "select * from model_login where upper(login) = upper('"+login+"')";
+
+        PreparedStatement statement = connection.prepareStatement(sql);
+
+        ResultSet resultado =  statement.executeQuery();
+
+        while (resultado.next()) /*Se tem resultado*/ {
+
+            modelLogin.setId(resultado.getLong("id"));
+            modelLogin.setEmail(resultado.getString("email"));
+            modelLogin.setLogin(resultado.getString("login"));
+            modelLogin.setSenha(resultado.getString("senha"));
+            modelLogin.setNome(resultado.getString("nome"));
+        }
+        return modelLogin;
+
+    }
+    public ModelLogin consultaUsuario(String login) throws Exception  {
+
+        ModelLogin modelLogin = new ModelLogin();
+
+        String sql = "select * from model_login where upper(login) = upper('"+login+"') and useradmin is false ";
+
+        PreparedStatement statement = connection.prepareStatement(sql);
+
+        ResultSet resultado =  statement.executeQuery();
+
+        while (resultado.next()) /*Se tem resultado*/ {
+
+            modelLogin.setId(resultado.getLong("id"));
+            modelLogin.setEmail(resultado.getString("email"));
+            modelLogin.setLogin(resultado.getString("login"));
+            modelLogin.setSenha(resultado.getString("senha"));
+            modelLogin.setNome(resultado.getString("nome"));
+
+        }
+        return modelLogin;
+    }
+
+    public ModelLogin consultaUsuario(String login, Long userLogado) throws Exception  {
+
+        ModelLogin modelLogin = new ModelLogin();
+
+        String sql = "select * from model_login where upper(login) = upper('"+login+"') and useradmin is false and usuario_id = " + userLogado;
+
+        PreparedStatement statement = connection.prepareStatement(sql);
+
+        ResultSet resultado =  statement.executeQuery();
+
+        while (resultado.next()) /*Se tem resultado*/ {
+
+            modelLogin.setId(resultado.getLong("id"));
+            modelLogin.setEmail(resultado.getString("email"));
+            modelLogin.setLogin(resultado.getString("login"));
+            modelLogin.setSenha(resultado.getString("senha"));
+            modelLogin.setNome(resultado.getString("nome"));
+
+        }
+        return modelLogin;
+    }
+
+    public ModelLogin consultaUsuarioID(String id, Long userLogado) throws Exception  {
+
+        ModelLogin modelLogin = new ModelLogin();
+
+        String sql = "select * from model_login where id = ? and useradmin is false and usuario_id = ?";
 
         PreparedStatement statement = connection.prepareStatement(sql);
         statement.setLong(1, Long.parseLong(id));
-        return getModelLogin(modelLogin, statement);
-    }
+        statement.setLong(2, userLogado);
 
-    private ModelLogin getModelLogin(ModelLogin modelLogin, PreparedStatement statement) throws SQLException {
-        ResultSet resultado = statement.executeQuery();
+        ResultSet resultado =  statement.executeQuery();
 
-        while(resultado.next()) {
+        while (resultado.next()) /*Se tem resultado*/ {
+
             modelLogin.setId(resultado.getLong("id"));
             modelLogin.setEmail(resultado.getString("email"));
             modelLogin.setLogin(resultado.getString("login"));
@@ -114,24 +181,25 @@ public class DAOUsuarioRepository {
         }
         return modelLogin;
     }
+    public boolean validarLogin(String login) throws Exception {
+        String sql = "select count(1) > 0 as existe from model_login where upper(login) = upper('"+login+"');";
 
-    public boolean validaLogin(String login) throws Exception{
-        String sql = "SELECT COUNT(1) > 0 AS existe FROM model_login WHERE UPPER(login) = UPPER('"+login+"')";
         PreparedStatement statement = connection.prepareStatement(sql);
-        ResultSet resultado = statement.executeQuery();
 
-        if(resultado.next()) {
-            return resultado.getBoolean("existe");
-        }
+        ResultSet resutlado =  statement.executeQuery();
 
-        return false;
+        resutlado.next();/*Pra ele entrar nos resultados do sql*/
+        return resutlado.getBoolean("existe");
     }
+    public void deletarUser(String idUser) throws Exception {
+        String sql = "DELETE FROM model_login WHERE id = ? and useradmin is false;";
 
-    public void deletarUser(String idUser) throws Exception{
-        String sql = "DELETE FROM model_login WHERE id = ?";
         PreparedStatement prepareSql = connection.prepareStatement(sql);
+
         prepareSql.setLong(1, Long.parseLong(idUser));
         prepareSql.executeUpdate();
+
         connection.commit();
+
     }
 }
